Fix problems updating texture size. Also see bug #10456.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2984 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/display3d.c
@@ -940,7 +940,7 @@ static void flipsAndRots(unsigned int tileNumber, unsigned int i, unsigned int j
 	/* Used to calculate texture coordinates, which are 0-255 in value */
 	const float xMult = 1.0f / TILES_IN_PAGE_COLUMN;
 	const float yMult = 1.0f / TILES_IN_PAGE_ROW;
-	const float one = 1.0f / (TILES_IN_PAGE_COLUMN * getTextureSize());
+	const float one = 1.0f / (TILES_IN_PAGE_COLUMN * (float)getTextureSize());
 
 	/*
 	 * Points for flipping the texture around if the tile is flipped or rotated
@@ -4051,7 +4051,7 @@ static void drawTerrainWaterTile(UDWORD i, UDWORD j)
 		/* Used to calculate texture coordinates, which are 0-255 in value */
 		const float xMult = 1.0f / TILES_IN_PAGE_COLUMN;
 		const float yMult = 1.0f / (2.0f * TILES_IN_PAGE_ROW);
-		const float one = 1.0f / (TILES_IN_PAGE_COLUMN * getTextureSize());
+		const float one = 1.0f / (TILES_IN_PAGE_COLUMN * (float)getTextureSize());
 		const unsigned int tileNumber = getWaterTileNum();
 		TERRAIN_VERTEX vertices[3];
